RDBC-768 IndexesFromClientTest::canExecuteManyIndexes
