+++ ravendb/tests/jvm_migrated_tests/client_tests/indexing_tests/test_java_script_index.py
@@ -3,7 +3,7 @@ from typing import List
 from ravendb import IndexFieldOptions
 from ravendb.primitives import constants
 from ravendb.documents.indexes.definitions import FieldIndexing
-from ravendb.documents.indexes.index_creation import AbstractJavaScriptIndexCreationTask
+from ravendb.documents.indexes.abstract_index_creation_tasks import AbstractJavaScriptIndexCreationTask
 from ravendb.infrastructure.entities import User
 from ravendb.tests.test_base import TestBase
