+++ ravendb/tests/jvm_migrated_tests/issues_tests/test_ravenDB_13682.py
@@ -1,4 +1,4 @@
-from ravendb.documents.indexes.index_creation import AbstractIndexCreationTask
+from ravendb.documents.indexes.abstract_index_creation_tasks import AbstractIndexCreationTask
 from ravendb.documents.queries.spatial import PointField
 from ravendb.tests.test_base import TestBase
