+++ ravendb/tests/jvm_migrated_tests/issues_tests/test_ravenDB_6558.py
@@ -2,7 +2,7 @@ from typing import List, Optional
 
 from ravendb import HighlightingOptions
 from ravendb.documents.indexes.definitions import FieldIndexing, FieldStorage, FieldTermVector
-from ravendb.documents.indexes.index_creation import AbstractMultiMapIndexCreationTask
+from ravendb.documents.indexes.abstract_index_creation_tasks import AbstractMultiMapIndexCreationTask
 from ravendb.documents.queries.highlighting import Highlightings
 from ravendb.tests.test_base import TestBase
