+++ ravendb/tests/jvm_migrated_tests/mailing_list_tests/test_lazy_aggregation_embedded.py
@@ -1,4 +1,4 @@
-from ravendb.documents.indexes.index_creation import AbstractIndexCreationTask
+from ravendb.documents.indexes.abstract_index_creation_tasks import AbstractIndexCreationTask
 from ravendb.tests.test_base import TestBase
